Add ability to generalize over a HTTP request being async or not.
Move AsyncHttp to CallbackfulHttp, and instead of inheriting from
ThreadPoolExecutor, take any Executor as a parameter. Then, have
both AsyncHttp and SyncHttp inheriting from it. AsyncHttp injects a
ThreadPoolExecutor, as before, while SyncHttp injects an Executor
that just executes Runnables right away, in the current thread.
Thus, SyncHttp exposes an “asynchronous” (callbackful) interface
while behaving synchronously.

By using CallbackfulHttp, we can abstract over whether an operation
is executed synchronously or asynchronously.

To make this more user-friendly, AnySyncHttp wraps both a SyncHttp
and an AsyncHttp, and lets you specify how to execute (class Execute) a
request (class Request) and then getting its result synchronously or
asynchronously (in another thread), as a method call result or passed
to a callback, with the `sync` and `async` methods respectively.

This design should make the Http class obsolete, once/if all the code
that uses it is ported. It could also be left as a proxy to SyncHttp.

(I did this after I renamed something in a `foo` method but forgot to
do the same in a `fooAsync` method, causing a bug that was reported by
a user.)
